feat: 自己看提交了啥（反正上一次提交的坑没填qwq

+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/Origin.kt
@@ -12,7 +12,7 @@ data class Origin(
     val pre_dy_id_str: String,
     val r_type: Int?,
     val repost: Int,
-    val rid: Int,
+    val rid: Long,
     val rid_str: String,
     val status: Int,
     val stype: Int?,
